test: use pytest
